+++ Sorting Algorithms/InsertionSort.cpp
@@ -1,34 +1,39 @@
+//C++ program to implement Insertion Sort
 #include <iostream>
-#include <string>
-#include <math.h>
-#include <algorithm>
-#include <vector>
-#include <stack>
-#include <queue>
-#include <set>
-#include <map>
-#include <utility>
 using namespace std;
-typedef long long ll;
-using namespace std;
-void insertionSort(int arr[], int n){
-    int i,key,j;
-    for(i=1;i<n;i++){
-        key=arr[i];
-        j=i-1;
+
+//Funtion performing Insertion Sort
+void InsertionSort(int *arr, int size){
+    int i, j, key;
+    for(i=1; i<size; i++){
+        key = arr[i];
+        j = i-1;
         while(j>=0 && arr[j]>key){
-            arr[j+1]=arr[j];
-            j=j-1;
+            arr[j+1] = arr[j];
+            j-=1;
         }
-        arr[j+1]=key;
+        arr[j+1] = key;
     }
 }
+
+//Funtion displaying the Sorted Array
+void Display(int *arr, int size){
+    cout<<"Sorted Array: ";
+    int i=0;
+    while(i<size){
+        cout<<arr[i++]<<" ";
+    }
+}
+
 int main(){
-    int arr[] ={2,1,3,5,6};
-    int i,n=sizeof(arr)/sizeof(arr[0]);
-    insertionSort(arr,n);
-    for(i=0;i<n;i++)
-        cout << arr[i] << " ";
-    cout << endl;
+    int i, size;
+    cout<<"Enter the size: "; cin>>size;
+    int *arr = new int[size];
+    cout<<"Enter "<<size<<" elements: ";
+    for(i=0; i<size; i++){
+        cin>>arr[i];
+    }
+    InsertionSort(arr,size);
+    Display(arr,size);
     return 0;
 }
